#MC-349 Add Work Development 4h add new Element models to artefact and datamigration from Metal
As a first step before extracting information from stratigraphy and in order to avoid duplicates and manual entry errors found in Metal records
we create a new django Element model containing all Elements from the periodic table.
This is based on and duplicate of the neo4j Element used in stratigraphy, as we don't have django/neo4j "foreign-key" support.

Then we convert all references to Metal to Element ones:
  - metal1,metalx (using Metal) ->  metal_e_1 and metal_e_x (using Element)

Both new and old fields are kept and displayed for review in this commit